+++ build.sbt
@@ -227,8 +227,14 @@ lazy val `layers-accumulo` = project
   .settings(commonSettings)
   .settings(Settings.`layers-accumulo`)
 
+lazy val `layers-cassandra` = project
+  .dependsOn(layers)
+  .settings(commonSettings)
+  .settings(Settings.`layers-cassandra`)
+
 lazy val cassandra = project
   .dependsOn(
+    `layers-cassandra`,
     spark % "compile->compile;test->test", // <-- spark-testkit update should simplify this
     `spark-testkit` % Test
   )
