--- cassandra/src/main/resources/META-INF/services/geotrellis.layers.io.ValueReaderProvider
@@ -1,1 +1,0 @@
-geotrellis.spark.io.cassandra.CassandraLayerProvider
