--- cassandra/src/main/resources/META-INF/services/geotrellis.spark.io.AttributeStoreProvider
@@ -1,1 +1,0 @@
-geotrellis.spark.io.cassandra.CassandraLayerProvider
