+++ cassandra/src/main/resources/META-INF/services/geotrellis.spark.io.LayerReaderProvider
@@ -1,1 +1,1 @@
-geotrellis.spark.io.cassandra.CassandraLayerProvider
+geotrellis.spark.io.cassandra.CassandraSparkLayerProvider
