+++ cassandra/src/main/scala/geotrellis/spark/io/cassandra/CassandraLayerCopier.scala
@@ -16,13 +16,15 @@
 
 package geotrellis.spark.io.cassandra
 
-import geotrellis.layers.LayerId
 import geotrellis.tiling.{Boundable, Bounds}
-import geotrellis.spark._
+import geotrellis.layers._
+import geotrellis.layers.cassandra._
+import geotrellis.layers.avro._
 import geotrellis.spark.io._
-import geotrellis.layers.io.avro._
 import geotrellis.util._
+
 import org.apache.spark.SparkContext
+
 import spray.json.JsonFormat
 
 import scala.reflect.ClassTag
