+++ cassandra/src/main/scala/geotrellis/spark/io/cassandra/CassandraLayerManager.scala
@@ -16,14 +16,16 @@
 
 package geotrellis.spark.io.cassandra
 
-import geotrellis.layers.LayerId
 import geotrellis.tiling.{Boundable, Bounds}
-import geotrellis.spark._
+import geotrellis.layers._
+import geotrellis.layers.cassandra._
+import geotrellis.layers.avro.AvroRecordCodec
+import geotrellis.layers.index._
 import geotrellis.spark.io._
-import geotrellis.layers.io.avro.AvroRecordCodec
-import geotrellis.layers.io.index._
 import geotrellis.util._
+
 import org.apache.spark.SparkContext
+
 import spray.json.JsonFormat
 
 import scala.reflect.ClassTag
