+++ cassandra/src/main/scala/geotrellis/spark/io/cassandra/CassandraLayerReader.scala
@@ -16,13 +16,16 @@
 
 package geotrellis.spark.io.cassandra
 
-import geotrellis.layers.LayerId
 import geotrellis.tiling.{Boundable, Bounds, EmptyBounds, KeyBounds}
-import geotrellis.spark._
-import geotrellis.spark.io._
-import geotrellis.layers.io.avro._
+import geotrellis.layers._
+import geotrellis.layers.avro._
+import geotrellis.layers.cassandra._
+import geotrellis.spark.ContextRDD
+import geotrellis.spark.io.FilteringLayerReader
 import geotrellis.util._
+
 import org.apache.spark.SparkContext
+
 import spray.json._
 
 import scala.reflect._
