+++ cassandra/src/main/scala/geotrellis/spark/io/cassandra/CassandraRDDWriter.scala
@@ -16,25 +16,29 @@
 
 package geotrellis.spark.io.cassandra
 
+import geotrellis.layers._
+import geotrellis.layers.avro._
+import geotrellis.layers.avro.codecs._
+import geotrellis.layers.cassandra._
+import geotrellis.layers.cassandra.conf.CassandraConfig
 import geotrellis.spark.io._
-import geotrellis.layers.io.avro._
-import geotrellis.layers.io.avro.codecs._
-import geotrellis.spark.io.cassandra.conf.CassandraConfig
 import geotrellis.spark.util.KryoWrapper
+
 import com.datastax.driver.core.DataType._
 import com.datastax.driver.core.querybuilder.QueryBuilder
 import com.datastax.driver.core.querybuilder.QueryBuilder.{eq => eqs}
 import com.datastax.driver.core.ResultSet
 import com.datastax.driver.core.schemabuilder.SchemaBuilder
+
 import cats.effect.IO
 import cats.syntax.apply._
+
 import org.apache.avro.Schema
 import org.apache.spark.rdd.RDD
+
 import java.nio.ByteBuffer
 import java.util.concurrent.Executors
 import java.math.BigInteger
-
-import geotrellis.layers.LayerId
 
 import scala.concurrent.ExecutionContext
 import scala.collection.JavaConverters._
