+++ layers-cassandra/src/main/resources/reference.conf
@@ -22,9 +22,6 @@ geotrellis.cassandra {
   usedHostsPerRemoteDc = 0
   allowRemoteDCsForLocalConsistencyLevel = false
   threads {
-    rdd {
-      write = default
-      read  = default
-    }
+    collection.read = default
   }
 }
