+++ layers-cassandra/src/main/scala/geotrellis/layers/cassandra/CassandraCollectionLayerReader.scala
@@ -14,14 +14,13 @@
  * limitations under the License.
  */
 
-package geotrellis.spark.io.cassandra
+package geotrellis.layers.cassandra
 
-import geotrellis.layers.{ContextCollection, LayerId}
 import geotrellis.tiling.{Boundable, Bounds, EmptyBounds, KeyBounds}
-import geotrellis.spark._
-import geotrellis.spark.io._
-import geotrellis.layers.io.avro._
+import geotrellis.layers._
+import geotrellis.layers.avro._
 import geotrellis.util._
+
 import spray.json._
 
 import scala.reflect._
