+++ layers-cassandra/src/main/scala/geotrellis/layers/cassandra/CassandraLayerHeader.scala
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package geotrellis.spark.io.cassandra
+package geotrellis.layers.cassandra
 
-import geotrellis.spark.io.{LayerHeader, LayerType, AvroLayerType}
+import geotrellis.layers.{LayerHeader, LayerType, AvroLayerType}
 
 import spray.json._
